backend/requirements.txt
# Core Framework
Flask>=2.3.0
flask-cors>=4.0.0
Werkzeug>=2.3.0

# Machine Learning
lightgbm>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Network & HTTP
requests>=2.31.0
urllib3>=2.0.0

# Email Processing
python-dateutil>=2.8.0

# Domain Analysis
python-whois>=0.8.0

# Database
# SQLAlchemy>=2.0.0  # Optional ORM support

# Text Processing
textblob>=0.17.0

# Production Server
waitress>=2.1.0
gunicorn>=21.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
